requirements.txt
anthropic>=0.25.0
requests>=2.31.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
jinja2>=3.1.0
